refactor: molecules folder정리
