+++ src/components/molecules/Home/HomeMoviesFetchBoard/index.js
@@ -1,8 +1,8 @@
 import { Button, Typography } from "@mui/material"
 import RefreshIcon from "@mui/icons-material/Refresh"
 import { useSetRecoilState } from "recoil"
-import { moviesFetchTriggerAtom } from "../../../store/moviesFetchTriggerAtom"
-import HomeBoardBox from "../../atoms/HomeBoardBox"
+import { moviesFetchTriggerAtom } from "../../../../store/moviesFetchTriggerAtom"
+import HomeBoardBox from "../../../atoms/HomeBoardBox"
 
 const HomeMoviesFetchBoard = ({ titleText }) => {
   const setIsFetchRequired = useSetRecoilState(moviesFetchTriggerAtom)
