+++ src/pages/Todos/index.js
@@ -4,7 +4,7 @@ import Label from "../../components/atoms/Label"
 import NormalButton from "../../components/atoms/NormalButton"
 import { useTranslation } from "react-i18next"
 import TodosList from "../../components/organisms/TodosList"
-import TodosCounter from "../../components/molecules/TodosCounter"
+import TodosCounter from "../../components/molecules/Todos/TodosCounter"
 import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil"
 import { todoCountAtom } from "../../store/todoCountAtom"
 import { Box } from "@mui/material"
